[fork] More Clang stdlib double/float function replacement

+++ src/screens/gm/gameMasterScreen.cpp
@@ -612,7 +612,7 @@ void GameMasterScreen::update(float delta)
             engine->setGameSpeed(
                 game_speed > 0.0f
                     ? 0.0f
-                    : std::powf(2.0f, static_cast<float>(game_time_scale->getSelectionIndex()))
+                    : std::pow(2.0f, static_cast<float>(game_time_scale->getSelectionIndex()))
             );
     }
